Except slice test things and split HGCalSimTest into MuonDetHitAnalyser and MuonTrackAnalyser
