+++ MuonAnalyser/test/runHGCalSimTest.py
@@ -20,7 +20,8 @@ process.source = cms.Source("PoolSource", fileNames = cms.untracked.vstring())
 process.source.skipEvents = cms.untracked.uint32(0)
 
 #process.source.fileNames.append('/store/user/yekang/CRAB_PrivateMC/HGCalStainless_default_me0_singleNu_RECO/190129_065738/0000/singleNu_GEN-SIM-DIGI_10.root')
-process.source.fileNames.append('/store/user/yekang/CRAB_PrivateMC/HGCalStainless_default_me0_singlePion_RECO_v2/190213_061518/0000/singlePion_GEN-SIM-DIGI_10.root')
+#process.source.fileNames.append('/store/user/yekang/CRAB_PrivateMC/HGCalStainless_default_me0_singlePion_RECO_v2/190213_061518/0000/singlePion_GEN-SIM-DIGI_10.root')
+process.source.fileNames.append('/store/user/yekang/CRAB_PrivateMC/HGCalStainless_default_me0_singleNu_RECO/190129_065738/0000/singleNu_GEN-SIM-DIGI_10.root')
 #process.source.fileNames.append('/store/user/yekang/me0/tenMu_modified/tenMu_GEN-SIM-DIGI_050.root')
 #process.source.fileNames.append('/store/user/yekang/me0/tenMu_default/tenMu_GEN-SIM-DIGI_060.root')
 #from glob import glob
@@ -35,25 +36,29 @@ process.options = cms.untracked.PSet(
 )
 
 process.TFileService = cms.Service("TFileService",fileName = cms.string("histo.root"))
-process.HGCalSimTest = cms.EDAnalyzer('HGCalSimTest',
+process.MuonDetHitAnalyser = cms.EDAnalyzer('MuonDetHitAnalyser',
     me0Digis = cms.InputTag("simMuonME0Digis"),
-    me0Segments = cms.InputTag("me0Segments"),
     me0RecHits = cms.InputTag("me0RecHits"),
-    cscSegments = cms.InputTag("cscSegments"),
     csc2DRecHits = cms.InputTag("csc2DRecHits"),
     gemDigis = cms.InputTag("simMuonGEMDigis"),
-    gemSegments = cms.InputTag("gemSegments"),
     gemRecHits = cms.InputTag("gemRecHits"),
     dtDigis = cms.InputTag("simMuonDTDigis"),
-    dt4DSegments = cms.InputTag("dt4DSegments"),
     dtRecHits = cms.InputTag("dt1DRecHits"),
     rpcDigis = cms.InputTag("simMuonRPCDigis"),
     rpcRecHits = cms.InputTag("rpcRecHits"),
+)
+
+process.MuonTrackAnalyser = cms.EDAnalyzer('MuonTrackAnalyser',
+    me0Segments = cms.InputTag("me0Segments"),
+    cscSegments = cms.InputTag("cscSegments"),
+    gemSegments = cms.InputTag("gemSegments"),
+    dt4DSegments = cms.InputTag("dt4DSegments"),
     simLabel = cms.InputTag("mix","MergedTrackTruth"),
     muonLabel = cms.InputTag("muons"),
     primaryVertex = cms.InputTag('offlinePrimaryVertices'),
 
 )
-process.p = cms.Path(process.HGCalSimTest)
+
+process.p = cms.Path(process.MuonDetHitAnalyser + process.MuonTrackAnalyser)
 
 process.schedule = cms.Schedule(process.p)
